chore: Move InputView class to view folder
